Add FAB and more tests for FlashcardListActivity

+++ app/src/mock/java/com/gregspitz/flashcardappkotlin/data/source/FakeFlashcardLocalDataSource.kt
@@ -11,15 +11,27 @@ import com.gregspitz.flashcardappkotlin.data.model.Flashcard
 class FakeFlashcardLocalDataSource(context: Context) : FlashcardDataSource {
 
     private val database : MutableMap<String, Flashcard> = mutableMapOf()
+    private var failure = false
 
     companion object : SingletonHolder<FakeFlashcardLocalDataSource, Context>(
             ::FakeFlashcardLocalDataSource)
 
     override fun getFlashcards(callback: FlashcardDataSource.GetFlashcardsCallback) {
+        if (failure) {
+            callback.onDataNotAvailable()
+            return
+        }
+
         callback.onFlashcardsLoaded(database.values.toList())
+
     }
 
     override fun getFlashcard(flashcardId: String, callback: FlashcardDataSource.GetFlashcardCallback) {
+        if (failure) {
+            callback.onDataNotAvailable()
+            return
+        }
+
         val flashcard = database.get(flashcardId)
         if (flashcard == null) {
             callback.onDataNotAvailable()
@@ -29,12 +41,18 @@ class FakeFlashcardLocalDataSource(context: Context) : FlashcardDataSource {
     }
 
     override fun saveFlashcard(flashcard: Flashcard, callback: FlashcardDataSource.SaveFlashcardCallback) {
+        if (failure) {
+            callback.onSaveFailed()
+            return
+        }
+
         database.remove(flashcard.id)
         database[flashcard.id] = flashcard
     }
 
     override fun deleteAllFlashcards() {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+        database.clear()
+        failure = false
     }
 
     override fun refreshFlashcards() {
@@ -45,7 +63,7 @@ class FakeFlashcardLocalDataSource(context: Context) : FlashcardDataSource {
         flashcards.forEach { database[it.id] = it }
     }
 
-    fun clearFlashcards() {
-        database.clear()
+    fun setFailure(failure: Boolean) {
+        this.failure = failure
     }
 }
